% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{input}
\alias{input}
\title{Plik do wczytywania danych zewnetrznych koniecznych do dzialania pakietu}
\format{
Ramka danych zawiera:
\describe{}
}
\source{
Plik wskazniki pochodzi z EEA - Europejskiej Agenji Srodowiska
}
\usage{
input
}
\description{
Plik do wczytywania danych zewnetrznych koniecznych do dzialania pakietu
}
\examples{
\dontrun{
 plikdane
}
}
\keyword{datasets}
