% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obl_emisji.R
\name{obl_emisji}
\alias{obl_emisji}
\title{Funkcja obl_emisji}
\usage{
obl_emisji(
  dane = input,
  kategoria = "Passenger Cars",
  euro = "Euro 5",
  mode = "",
  substancja = c("CO", "EC")
)
}
\arguments{
\item{dane}{data_frame}

\item{kategoria}{char}

\item{euro}{char}

\item{mode}{char}

\item{substancja}{cahr}
}
\value{
data_frame
}
\description{
Funkcja obl_emisji
}
