% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiz_emisji.R
\name{wiz_emisji}
\alias{wiz_emisji}
\title{Title Funkcja wiz_emisji()}
\usage{
wiz_emisji(dane = podsum)
}
\arguments{
\item{dane}{data_frame}
}
\value{
column chart
}
\description{
Title Funkcja wiz_emisji()
}
\examples{
wiz_emisji()
}
